Update date formatting to match WP Settings (#150)
* Make dates on certificates follow date format from WP Settings

There is a backwards compatibility issue with the
`sensei_certificate_date_format` filter. Previously, for non-English
locales, the filter was expected to return a format to be used by
`strftime()` instead of `date()` or `date_i18n()`. Moving forward, the
filter should always return a format to be used by `date_i18n()`
regardless of the locale.

However, for backwards compatibility, we check for a `strftime()` format
string in non-English locales and use `strftime()` if needed.

* Add doc comment for filter